[parse] Move Span out of bve::parse::mesh to bve::parse
