+++ bve/src/parse/mesh/errors.rs
@@ -1,3 +1,5 @@
+use crate::parse::Span;
+
 #[derive(Debug, Clone, PartialEq)]
 pub struct MeshError {
     pub kind: MeshErrorKind,
@@ -90,20 +92,6 @@ impl From<csv::Error> for MeshError {
                 kind: MeshErrorKind::UnknownCSV,
                 span: e.position().into(),
             },
-        }
-    }
-}
-
-#[derive(Copy, Clone, Debug, PartialEq)]
-pub struct Span {
-    pub line: Option<u64>,
-}
-
-impl<'a> From<Option<&'a csv::Position>> for Span {
-    #[must_use]
-    fn from(p: Option<&'a csv::Position>) -> Self {
-        Self {
-            line: p.map(csv::Position::line),
         }
     }
 }
