+++ bve/src/parse/mod.rs
@@ -1,4 +1,6 @@
-//! All code for parsing and emitting files.
+//! File parsers, linters, and code generators
 
 pub mod mesh;
 mod util;
+
+pub use util::Span;
